+++ tests/test_distance.py
@@ -651,6 +651,42 @@ class TestDistance(util.ColorAssertsPyTest):
         'color1,color2,value',
         [
             ('red', 'red', 0),
+            ('red', 'orange', 0.3429),
+            ('red', 'yellow', 0.3841),
+            ('red', 'green', 0.3872),
+            ('red', 'blue', 0.361),
+            ('red', 'indigo', 0.333),
+            ('red', 'violet', 0.3252),
+            ('red', 'white', 0.3464),
+            ('red', 'black', 0.3852),
+            ('red', 'gray', 0.3266),
+            ('red', 'red', 0),
+            ('orange', 'red', 0.3429),
+            ('yellow', 'red', 0.3841),
+            ('green', 'red', 0.3872),
+            ('blue', 'red', 0.361),
+            ('indigo', 'red', 0.333),
+            ('violet', 'red', 0.3252),
+            ('white', 'red', 0.3464),
+            ('black', 'red', 0.3852),
+            ('gray', 'red', 0.3266)
+        ]
+    )
+    def test_delta_e_helmlab(self, color1, color2, value):
+        """Test delta e Helmlab."""
+
+        print('color1: ', color1)
+        print('color2: ', color2)
+        self.assertCompare(
+            Color(color1).delta_e(color2, method="helmlab"),
+            value,
+            rounding=4
+        )
+
+    @pytest.mark.parametrize(
+        'color1,color2,value',
+        [
+            ('red', 'red', 0),
             ('red', 'orange', 37.32),
             ('red', 'yellow', 66.5558),
             ('red', 'green', 68.7175),
@@ -672,6 +708,7 @@ class TestDistance(util.ColorAssertsPyTest):
             ('gray', 'red', 47.5086)
         ]
     )
+
     def test_delta_e_cam16(self, color1, color2, value):
         """Test delta e CAM16 UCS."""
 
@@ -962,4 +999,3 @@ class TestClosest(util.ColorAsserts, unittest.TestCase):
 
         with self.assertRaises(ValueError):
             Color('red').delta_e('blue', method='cam02', space='lab')
-